test(eval): relocate database blank-semantics tests and pin blank-only DGET match (#281)
