+++ crates/formualizer-eval/src/engine/tests/mod.rs
@@ -7,6 +7,7 @@ mod clock_snapshot;
 mod common;
 mod cross_sheet_named_range_first_cell;
 mod cycle_detection;
+mod database_blank_semantics;
 mod deferred_dirty;
 mod demand_driven;
 mod dependency;
